appscript0004/appscript.py
import os
from googleapiclient.discovery import build
from google.oauth2 import service_account

class AppScript:
    def __init__(self):
        self.service = build('script', 'v1', credentials=self.get_credentials())

    def get_credentials(self):
        # Implement logic to get credentials
        pass

    def create_program(self, script_id):
        # Implement logic to create program
        pass

    def deploy_script(self, script_id):
        # Implement logic to deploy script
        pass