appscript0004/google_lens_db.py
import mysql.connector

class GoogleLensDB:
    def __init__(self):
        self.conn = mysql.connector.connect(
            user='username',
            password='password',
            host='localhost',
            database='google_lens_db'
        )
        self.cursor = self.conn.cursor()

    def get_data_from_rakuten(self):
        # Implement logic to get data from Rakuten
        pass

    def create_advertisement(self, product):
        # Implement logic to create advertisement
        pass

    def create_se(self, product):
        # Implement logic to create SE
        pass

    def create_tags(self, product):
        # Implement logic to create 20 tags
        pass

    def create_html(self, product):
        # Implement logic to create HTML
        pass

    def create_flex(self, product):
        # Implement logic to create FLEX
        pass